requirements.txt
# MCP Framework
mcp-use>=0.1.0

# Web UI
gradio>=4.0.0

# LLM Providers (install what you need)
langchain>=0.3.0
langchain-openai>=0.2.0
langchain-anthropic>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0

# Database
psycopg2-binary>=2.9.0

# Utilities
python-dotenv>=1.0.0
aiohttp>=3.9.0

# Optional: Observability
# langfuse>=2.0.0

# Optional: Local models
# langchain-ollama>=0.1.0
